(new play) Hints! Gifts! Lava, galore!
I've added signs to help the new Chunker along, especially if they didn't read the instructions.

I've also spread much of the rewards across a couple more chests. There's one more to come, as soon
as I can decide *what* to put in it. ;)

Finally, I moved the lava. It was way, way too easy to get too; and also really shocking to the new
Chunker. Now you've gotta work up to it; and it's a bit more obvious, with obsidian on the wall.